backend/requirements.txt
# SOVREN AI - Complete Production Requirements
# Enterprise-grade Digital Chief of Staff Platform

# ============================================
# CORE DEPENDENCIES
# ============================================

# Python version requirement
# Requires Python 3.12+

# Core ML/Audio Processing
numpy>=1.24.0,<2.0.0
torch>=2.0.0,<3.0.0
torchaudio>=2.0.0,<3.0.0
soundfile>=0.12.1,<1.0.0
librosa>=0.10.0,<1.0.0
sounddevice>=0.4.6,<1.0.0
scipy>=1.11.0,<2.0.0

# TTS and Speech Synthesis
styletts2>=1.0.0  # StyleTTS2 for high-quality speech synthesis
pyttsx3>=2.90,<3.0.0  # Fallback TTS engine
whisper>=1.0.0  # OpenAI Whisper for transcription

# Async and Networking
aiohttp>=3.9.0,<4.0.0
websockets>=12.0,<13.0.0
aiofiles>=23.0.0,<24.0.0
aiodns>=3.0.0,<4.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.13.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0  # PostgreSQL driver
redis>=5.0.0,<6.0.0  # Async Redis client
asyncpg>=0.27.0,<1.0.0

# API and Validation
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
pydantic>=2.0.0,<3.0.0
pyyaml>=6.0,<7.0.0

# Security
cryptography>=41.0.0,<42.0.0
python-jose>=3.3.0,<4.0.0  # JWT tokens
passlib[bcrypt]>=1.7.4,<2.0.0  # Password hashing
PyJWT>=2.8.0,<3.0.0

# Monitoring and Logging
prometheus-client>=0.18.0,<1.0.0
sentry-sdk>=1.35.0,<2.0.0
psutil>=5.9.0,<6.0.0  # System monitoring
structlog>=23.0.0,<24.0.0

# Data Processing
pandas>=1.3.0,<2.0.0
openpyxl>=3.0.0,<4.0.0  # Excel file support
scikit-learn>=1.0.0,<2.0.0

# HTTP Client
requests>=2.31.0,<3.0.0

# ============================================
# DEVELOPMENT DEPENDENCIES
# ============================================

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0

# Code Quality
black>=23.0.0,<24.0.0  # Code formatting
mypy>=1.5.0,<2.0.0  # Type checking
ruff>=0.1.0,<1.0.0  # Fast Python linter
flake8>=6.0.0,<7.0.0
bandit>=1.7.0,<2.0.0  # Security linting
safety>=2.0.0,<3.0.0  # Dependency security

# Documentation
sphinx>=7.0.0,<8.0.0
sphinx-rtd-theme>=1.3.0,<2.0.0

# Deployment
gunicorn>=21.0.0,<22.0.0
supervisor>=4.2.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0

# ============================================
# OPTIONAL DEPENDENCIES
# ============================================

# Optional but recommended
httpx>=0.25.0,<1.0.0  # Modern HTTP client
tenacity>=8.2.0,<9.0.0  # Retry logic
cachetools>=5.3.0,<6.0.0  # Caching utilities

# ============================================
# SYSTEM DEPENDENCIES (Install separately)
# ============================================

# These must be installed via system package manager:
# - postgresql-server
# - redis-server
# - nginx (for production)
# - ffmpeg (for audio processing)
# - sox (for audio manipulation)
# - portaudio (for audio I/O)
# - libasound2-dev (for ALSA audio)
# - libportaudio2 (for PortAudio)
# - libsndfile1 (for audio file I/O)
# - libfftw3-dev (for FFT)
# - libsamplerate0-dev (for audio resampling)
# - libavcodec-dev (for audio codecs)
# - libavformat-dev (for audio formats)
# - libavutil-dev (for audio utilities)
# - libswresample-dev (for audio resampling)
# - libavfilter-dev (for audio filters) 